finish first pass at refactor
